Half cleaned and fully cleaned floorplan prefabs
Added a super cleaned floorPlanCleaned prefab and a
floorplanHalfCleaned prefab, the full doesn’t look as good with the
neon outline shader. I still want to make a trim and see how that works
though. Test out the performance with this bad boy. Added colors for
each quadrant as well